Feature: add an option to clean_dump_archive or not (default is True)
git-svn-id: file:///home/code/svn/ovd/trunk@6852 2b2718be-6ee5-4ec5-ad1a-3fa14f9c7ca3

+++ OvdServer/ovd/Role/ApplicationServer/Role.py
@@ -95,7 +95,8 @@ class Role(AbstractRole):
 			Logger.error("Unable to purge group")
 			return False
 		
-		self.purgeArchives()
+		if Config.clean_dump_archive:
+			self.purgeArchives()
 		
 		if Config.multithread:
 			cpuInfos = Platform.System.getCPUInfos()
